requirements.txt
pandas
numpy
scikit-learn
xgboost
shap
streamlit
psycopg2-binary
joblib
plotly
